+++ package.json
@@ -6,12 +6,13 @@
     "scripts": {
         "test": "jest",
         "test:watch": "jest --watch",
-        "start": "start-storybook -p 9001 -c .storybook",
+        "start": "storybook dev -p 9001 -c .storybook",
         "lint": "tslint ./src/**/*.{ts,tsx}",
         "clean": "rm -rf ./dist ./node_modules",
         "build": "tsc",
         "format": "prettier --write \"**/*.{ts,tsx,json,md}\"",
-        "prepublishOnly": "rm -rf ./dist && npm test && npm run lint && npm run build"
+        "prepublishOnly": "rm -rf ./dist && npm test && npm run lint && npm run build",
+        "build-storybook": "storybook build"
     },
     "publishConfig": {
         "access": "public"
@@ -48,13 +49,16 @@
     },
     "devDependencies": {
         "@babel/core": "^7.21.0",
+        "@babel/preset-env": "^7.21.4",
+        "@babel/preset-react": "^7.18.6",
+        "@babel/preset-typescript": "^7.21.4",
         "@dump247/storybook-state": "^1.6.1",
-        "@storybook/addon-actions": "^6.5.16",
-        "@storybook/addon-essentials": "^6.5.16",
-        "@storybook/addon-storysource": "^6.5.16",
-        "@storybook/builder-webpack5": "^6.5.16",
-        "@storybook/manager-webpack5": "^6.5.16",
-        "@storybook/react": "^6.5.16",
+        "@storybook/addon-actions": "^7.0.6",
+        "@storybook/addon-essentials": "^7.0.6",
+        "@storybook/addon-mdx-gfm": "^7.0.6",
+        "@storybook/addon-storysource": "^7.0.6",
+        "@storybook/react": "^7.0.6",
+        "@storybook/react-webpack5": "^7.0.6",
         "@testing-library/jest-dom": "^5.16.5",
         "@testing-library/react": "^14.0.0",
         "@types/jest": "^29.4.0",
@@ -75,6 +79,7 @@
         "react": "^18.2.0",
         "react-dom": "^18.2.0",
         "react-leaflet": "^4.2.0",
+        "storybook": "^7.0.6",
         "styled-components": "^5.3.6",
         "ts-jest": "^29.0.5",
         "tslint": "^6.1.3",
